Implement generic env vars for queries.
